Add files via upload
This script is a handy tool for anyone who wants to organize and keep track of their PDF and EPUB ebooks.  Imagine you have a big folder on your computer (or maybe several folders) full of these ebooks, and it's getting hard to find what you're looking for.  This script will go through those folders (and any folders inside them), find all the PDF and EPUB files, and create a neat list for you.

Think of it like creating a catalog of your ebooks.  The list will show you:

Which folder each book is in.
The name of the book file.
How big the file is (so you know which ones might take up a lot of space).
The exact location of the file on your computer.
Once the script has made this list, it saves it into a file that you can open in a spreadsheet program (like Excel or Google Sheets). This makes it easy to sort your books, search for specific titles, or just get a good overview of your ebook collection.

You can tell the script which folders to look in, and it can even give the list file a random name so you don't accidentally overwrite old lists.  It's a great way to bring some order to your digital library!